Refactoring whole module. Sensors = Visual + Geometric. Used enumerations to differ. Minor corrections
