+++ flatland/agents/sensors/__init__.py
@@ -1,1 +1,2 @@
-from .collection import *
+from .visual_sensors import *
+from .geometric_sensors import *
