+++ flatland/agents/sensors/geometric_sensors/__init__.py
@@ -1,0 +1,1 @@
+from .collection import *
