+++ flatland/agents/sensors/geometric_sensors/collection/lidar.py
@@ -1,28 +1,37 @@
+from flatland.agents.sensors.sensor import *
+from flatland.agents.sensors.geometric_sensors.geometric_sensor import *
 from collections import defaultdict
 from pymunk.vec2d import Vec2d
 
-#Brait
-class LidarSensor():
+
+@SensorGenerator.register('lidar')
+class LidarSensor(GeometricSensor):
 
     def __init__(self, anatomy, custom_config):
 
         self.sensor_type = 'lidar'
+        
+        #Todo later: add default config, as in visual_sensors
         sensor_param = { **custom_config}
 
-        self.observation = None
+        super(LidarSensor, self).__init__(anatomy, sensor_param)
+
+
+        #Sensor paramters TODO: make it parametrable
+        self.FoV = 100 #in pixels
+        self.angles = [-180, -90, 0, 90, 180]
+        self.cones_number = len(self.angles)-1
+
+
 
         pass
 
 
     def update_sensor(self, current_agent, entities, agents):
 
-        #Sensor paramters TODO: make it parametrable
-        FoV = 100 #in pixels
-        angles = [-180, -90, 0, 90, 180]
-        cones_number = len(angles)-1
 
         #Initialising ouput
-        output = [dict() for i in range(cones_number)]
+        output = [dict() for i in range(self.cones_number)]
 
         #Current's agent Shape
         agent_shape = current_agent.frame.anatomy['base'].shape
@@ -65,7 +74,7 @@ class LidarSensor():
             #add here: Tests on entity_type : can the entity be detected ?
 
 
-            for i in range(cones_number):
+            for i in range(self.cones_number):
                 output[i][entity_type] = 1
 
             #For each entity
@@ -81,7 +90,7 @@ class LidarSensor():
                 #Distance check
                 distance = agent_position.get_distance(near_point)
 
-                if distance > FoV:
+                if distance > self.FoV:
                     print("Distance is >")
                     break
 
@@ -89,8 +98,8 @@ class LidarSensor():
                 angle = agent_position.get_angle_degrees_between(near_point) #in degrees
                 cone = None
 
-                for i in range(len(angles)-2):
-                    if angle >= angles[i] and angle < angles[i+1]:
+                for i in range(len(self.angles)-2):
+                    if angle >= self.angles[i] and angle < self.angles[i+1]:
                         cone = i
 
                 if cone is None:
@@ -104,7 +113,7 @@ class LidarSensor():
 
                     output[cone][entity_type] = 1
 
-                normalised_distance = distance/FoV
+                normalised_distance = distance/self.FoV
 
                 #Keeping only the nearest distance
                 if output[cone][entity_type] > normalised_distance:
