+++ flatland/agents/sensors/visual_sensors/collection/rgb_sensor.py
@@ -1,4 +1,4 @@
-from flatland.agents.sensors.sensor import SensorGenerator, Sensor
+from flatland.agents.sensors.visual_sensors.visual_sensor import *
 import numpy as np
 import cv2
 import os, yaml
@@ -9,7 +9,7 @@ with open(os.path.join(__location__, 'sensor_default.yml'), 'r') as yaml_file:
 
 
 @SensorGenerator.register('rgb')
-class RgbSensor(Sensor):
+class RgbSensor(VisualSensor):
 
     def __init__(self, anatomy, sensor_param):
