+++ flatland/agents/sensors/visual_sensors/collection/sensor_default.yml
